requirements.txt
fastapi==0.109.0
motor==3.3.2
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pydantic
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
certifi==2024.2.2
email-validator==2.1.0.post1
aiosmtplib
sendgrid>=6.9.0
gunicorn==21.2.0
prometheus-client==0.19.0
sentry-sdk==1.40.0
python-json-logger==2.0.7 